+++ CumulusMX/JsonStationMqtt.cs
@@ -3,7 +3,6 @@ using System.Threading;
 using System.Threading.Tasks;
 
 using MQTTnet;
-using MQTTnet.Client;
 
 
 namespace CumulusMX
@@ -20,7 +19,7 @@ namespace CumulusMX
 			cumulus = cmx;
 			station = stn;
 
-			var mqttFactory = new MqttFactory();
+			var mqttFactory = new MqttClientFactory();
 
 			/*
 			var protocolType = cumulus.MQTT.IpVersion switch
@@ -100,7 +99,7 @@ namespace CumulusMX
 		}
 
 
-		private static async Task Connect(MQTTnet.Client.MqttClientOptions options)
+		private static async Task Connect(MQTTnet.MqttClientOptions options)
 		{
 			try
 			{